LayoutTests/media/video-src-change-expected.txt

Test that changing src attribute triggers load when network state is empty.
Test that it does not trigger load when network state is non-empty.
EVENT(emptied)
EXPECTED (video.networkState == '0') OK
EVENT(loadstart)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
EVENT(load)
EXPECTED (video.networkState != '0') OK
END OF TEST
